+++ src/Raphael/bench.h
@@ -1,0 +1,11 @@
+#include <Raphael/Raphael.h>
+
+
+
+namespace raphael::bench {
+/** Runs the benchmark
+ *
+ * \param engine engine to benchmark
+ */
+void run(Raphael& engine);
+}  // namespace raphael::bench
