+++ uci.cpp
@@ -1,6 +1,6 @@
 #include <GameEngine/consts.h>
 #include <Raphael/Raphael.h>
-#include <Raphael/tests.h>
+#include <Raphael/bench.h>
 #include <Raphael/tunable.h>
 
 #include <condition_variable>
@@ -194,9 +194,6 @@ int main(int argc, char** argv) {
             lock_guard<mutex> engine_lock(engine_mutex);
             raphael::bench::run(engine);
             return 0;
-        } else if (!strcmp(argv[1], "test")) {
-            raphael::test::run_all(false);
-            return 0;
         }
         lock_guard<mutex> lock(cout_mutex);
         cout << "info string ignoring unknown command line arguments\n" << flush;
@@ -247,9 +244,6 @@ int main(int argc, char** argv) {
 #else
             cout << "info string this is not a tunable build\n" << flush;
 #endif
-
-        } else if (uci_command == "test") {
-            raphael::test::run_all(false);
 
         } else if (uci_command == "bench") {
             lock_guard<mutex> engine_lock(engine_mutex);
